backend/requirements.txt
# CapeTown GIS Hub — Python Backend Production Dependencies
# Install: pip install -r requirements.txt
# Note: GDAL is NOT installed via pip — it comes from the OSGEO Docker base image

# API framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.3.0
httpx>=0.27.0

# Database — async path
sqlalchemy[asyncio]>=2.0.30
geoalchemy2>=0.15.0
asyncpg>=0.29.0

# Database — sync path (GeoPandas operations)
psycopg2-binary>=2.9.9

# GIS core
# GDAL is provided by the OSGEO Docker base image — never install it via pip
# See GOTCHA-PY-001 in docs/gotchas.md
geopandas>=0.14.4
fiona>=1.9.6
shapely>=2.0.4
pyproj>=3.6.1
pyogrio>=0.9.0
rasterio>=1.3.10
rio-cogeo>=5.3.0

# GIS format-specific
laspy>=2.5.4
ezdxf>=1.3.4
arcgis>=2.3.0

# Raster / satellite
rioxarray>=0.17.0
xarray>=2024.6.0
earthengine-api>=0.1.400
sentinelsat>=1.2.1

# OGC services
pygeoapi>=0.18.0

# Celery + Redis
celery[redis]>=5.4.0
redis>=5.0.7
flower>=2.0.1

# Storage
boto3>=1.34.0
python-magic>=0.4.27

# Auth — Supabase JWT validation via JWKS
python-jose[cryptography]>=3.3.0

# Observability
sentry-sdk[fastapi]>=2.7.0
structlog>=24.2.0
